Avoid some allocations in ABISig::from_func_sig
In cranelift_codegen::machinst::abi_impl, the function
`ensure_struct_return_ptr_is_returned` sometimes, but not always, needs
to modify the `ir::Signature` to add a struct return. This commit makes
it return a `Cow<ir::Signature>` so that if it's unmodified, it doesn't
have to clone the signature.

That wouldn't be useful if all callers wanted an owned copy of the
signature, as it turns out one of them do. But three of the four callers
do not; they're happy with a borrow.

DHAT reports that when compiling the Sightglass Spidermonkey benchmark,
there were about 44k allocations in this particular call stack,
averaging 12 bytes, and with an average lifetime of about 2700
instructions.

On that benchmark, either this patch eliminates all of those
allocations, or it has hidden them from me somehow.