Provide ValueRegs::from_iter
Of the three existing callers of ValueRegs::two(), two of them were
checking the length of some slice to decide whether to call that or
ValueRegs::one(). That meant the ValueRegs abstraction leaked: users had
to know how many physical registers a value could be split across.

Implementing FromIterator enables clearer code at its call-sites and
centralizes the details of how many registers can fit in a ValueRegs.

I left all the uses of ValueRegs::one() alone. It's pretty common to
know ahead of time that one register suffices, so that's a handy
shorthand.