Require a Default impl for SecondaryMap values
All the existing uses of SecondaryMap's support for custom defaults
actually use values that are reasonable defaults for their types. So
just ensure those types have a reasonable implementation of Default and
stop using SecondaryMap::with_default.

The one place where this changes behavior is in the serde implementation
for SecondaryMap itself, which previously serialized the default value
before the elements of the map. The default value is now always known at
the type level, and I'm removing support for specifying a default when
constructing a map. So I've removed the extra element from the
serialized representation.