cranelift-codegen: Return CompiledCode by value
One reason to store `CompiledCode` in a cranelift-codegen `Context` is
so that a series of method calls on Context can all access the same
compilation result. However, functions that read the `compiled_code`
field didn't access any other field of `Context`, so I moved those
accessors to `CompiledCode` in a previous PR.

Another reason to keep the `CompiledCode` in the `Context` would be to
reuse its heap allocations. However, the existing code didn't do that:
`Context::clear()` assigned `None` to the `compiled_code` field, which
dropped any allocations that might have been in there before reusing the
context.

So this commit removes the `Context::compiled_code` field, and changes
the various functions that produce a `CompiledCode` to move ownership to
the caller, instead of borrowing the inside of the `Context`.

This statically eliminates a potential source of errors (trying to
access the compiled code without first calling compile), makes the
borrow-checker happier in several cases, and allows callers to avoid
copying if they want.