s390x: Simplify `add_trap` flag
On s390x, the `mem_*emit` functions currently all take a boolean flag
named `add_trap`. If this flag is false, then they do not actually
record trap metadata, even if the flags say the memory access can trap.

However, there is exactly one circumstance when this flag is false,
which happens when emitting a LoadAddr instruction. So rather than
passing an extra parameter that's almost always true, I changed that one
instruction to override the memory flags to indicate that accessing the
address can't trap.

There was one additional detail controlled by this flag. The
`have_unaligned_pcrel` field passed to `mem_finalize` is set to true
only for LoadAddr. The condition for recognizing that it's a LoadAddr
amounted to checking that `add_trap` was false. I've changed that to
check `opcode_ril` for the specific opcode of LoadAddr instead.