cranelift: VReg aliases must not change register class
I wanted it to be true that VReg aliases always have the same type in
vreg_types, but I discovered that is not true. (I think because of
bitcasts, mostly, but maybe there are other reasons.) Still, at least
the register class must match, or register allocation would be
impossible.