wasmtime: Generalize const-expression evaluator
When evaluating a constant expression, we may or may not have an
instance available in which to evaluate `global.get` and `ref.func`.
When we don't have an instance, there are still some expressions we can
const-eval. Allowing the caller to provide however much context they
actually have allows us to aggressively const-eval whenever possible.