cranelift-frontend: Put debug labels on variables, not values
We need to get the debug labels onto the SSA values eventually, but only
cranelift-frontend knows which values it introduces at phi join points,
so it should be responsible for applying the labels to the right values.

Currently, cranelift-wasm applies an additional debug label every time a
value is used, to avoid needing to know when block params are created.

But this creates a lot of duplicate labels if there are a lot of
`local.get` instructions for the same local without an intervening
`local.set`.