wasmtime: Refactor trap-handling
This commit groups together all the details about a trap that have to be
collected from a signal handler: the program counter, frame pointer, and
faulting address, as of the time that the trap occurred.

I also moved the call to set_jit_trap inside test_if_trap for every
platform that uses both methods. Only the implementation for Mach ports
still needs to call set_jit_trap because it doesn't use test_if_trap.

In addition I'm fixing an unrelated doc comment that I stumbled across
while working on this.